JIT: Fix BuiltinFunc not to inherit Expr class
